Add `initOnly` option and ability to manually load scripts
When `initOnly` is true, the registered script will only call the init function but will _not_ load the remote script.

To call the remote script, the `deferScript('load', { name: 'someRegisteredName’ })` must be executed.

This is useful if you want manual control over when the remote script is laoded (e.g. on scroll or intersection)